Add L1RoutingGraph to eventually replace {Wafer,HICANN}Graph
Compared to {Wafer,HICANN}Graph, we adopt a wafer-local view of the L1
routing problem, i.e. HICANNGlobal → HICANNOnWafer.
L1RoutingGraph will provide methods to alter the connectivity of the
graph, e.g. iteratively adding new HICANNs or disabling certain
connections.  The API for this will be based on hardware coordinates.
As a consequence, calls to the resource manager etc. can be moved out of
the graph code, leading to better encapsulation and testability.

Change-Id: I5972e1ecb2f99d040857f88ac8bde7f1f30eb257