Refactor merger tree routing to split off configuration
This commit does not introduce functionality changes but splits off the
representation of the merger tree into a separate class MergerTreeGraph,
which is in turn used in MergerTreeRouter to find a mapping from neuron
blocks to DNC mergers.
This mapping can then be passed to MergerTreeConfigurator to configure
the appropriate sthal container.  While it is planned to decouple this
step from the mapping process, we call it immediately afterwards at the
moment to conserve the current behavior.

Change-Id: Ic4ee9e9cafba878f739df94af6075730e752dc4a