Add option to pass in sthal::Wafer via pynn.setup()
This configuration container can then be connected to the hardware once
and used with "leaner" configurators on subsequent calls to pynn.run().

Note that this mode of operation is only available when compiled with
"--without-ester", as due to happenstance[1] no serialization of Metadata
objects is performed in this case and marocco has direct access to the
data structures created by the user.
This will have to be formalized at a later point in time.  Ideally the
user should eventually have the means to create shared_ptr<T> for all
relevant objects (sthal::Wafer, marocco::results::Marocco), via factory
methods exposed as python ctors.  Then, those pointers could be passed
to the mapping directly and the marocco::Runtime object could be retired.

[1] see issue 2299

Change-Id: I805a68528dab2509f5169e96321303ee7d9484d9