Change input placement to enforce routing of BG events
As the L1 events necessary for locking of repeaters and synapse drivers
would not arrive early enough when sent via external input, we need to
merge background events generated on-chip until a different solution is
in place (e.g. using pbmem to send events before the experiment start).

Previously, external input could only be sent via DNCMergerOnHICANN(7)
using the `restrict_rightmost_neuron_blocks` hack, which had the effect
of blacklisting the corresponding neuron blocks and configuring their
background generators to be merged with the external events.

After this changeset, every DNC merger that is not used for hardware
neurons (the current implementation assumes that every neuron on chip
will need spike readout) is a viable candidate for placement of external
input, provided a background generator can be connected.

The merger routing result now consist of a full mapping of neuron blocks
to DNC mergers.

Change-Id: Ia2959e60ca239da7ec177e2c1f320ddd53967755