test-ManualPlacement.py: new test case for NeuroBlock-exhaustion
NeuronBlock-exhaustion behaves different from HICANN-exhaustion and should also be tested.

definition:
X-exhaustion: X cant hold the requested amount of Neurons.

As the automatic placer works NB-wise a test is needed to test correct placement of multiple NB.
The automatic placer is called for manual placement requests in L.272 of NeuronPlacement.cpp.

the current tests only test placment on 1 NB or HICANN-exhaustion.
1 * 4 = 4 < 64 = 1NB
2 * 4 = 8 < 1NB
10 * 4 = 40 < 1NB
999 * 4 = 3996 > 8 NB  = 1 HICANN

the new test uses multiple NB:
100* 4 = 400 > 64 = 1 NB   &&   400 < 8*64 = 512 = 8 NB = 1 HICANN

The if statement in L.21 is triggered from 128 Neurons, so the tests in L.31 were not executed with more than one NB.
to reach the tets in L.31 a Population size between 17 and 127 Neurons is needed.

So this change helps in debugging of placement strategies by requiring at least 2 NB  but at most 1 HICANN (8 NB).

Change-Id: I55a63377a52230a0dac6976ea1ab792fab7c4904