Class Wrapping for Population placement.
A new namespace for user defined algorithms is defined:
marocco::placement::algorithms

Placement strategies can be created as classes inside this namespace
inherited from ::PlacePopulationsBase.

An instances of the chosen placement class can be passed as a parameter to pymarocco.
Settings to influence the behaviour of the placement strategy shall be set in the instances of the class.

The classes are exposed to python to create and modify them by the user.
This required also some changes to included classes.

Sample strategies are:
• byNeuronBlockEnumAndPopulationID: places the populations linear on the
  NeuronBlocks
• bySmallerNeuronBlockAndPopulationID: the default placement strategy used
  so far; placing populations linear on a torn apart "spiral".

Change-Id: I996e7ae08aad5c1d3882a7b8b9c6b108d4522e33