Split external spike sources if driver requirements exceeds hardware maximum
Upon placement of external spike sources the maximum driver chain length is calculated.
If it exceeds the hardware settings, the placement is undone, split and queued again.
Splitting the spike source is a way to reduce the number of required synapse drivers
to keep it within hardware specifications.
A logarithmic back-off is used by splitting into two equal parts.

Change-Id: I994dd6f07995be922fe2153dab2b19bbe9dd6b16