Fix throw/potential segfault (in manual placement)
Previously, the code threw an "key does not exist in map" error or,
worse, segfaulted when trying to access an non-existing HICANN in the
resource manager data structure. However, non-existing redman resources
are a valid state when comissioning a system (or using a 'patchy'
single-chip system). Therefore, we have to actively check for resource
existance when using user-specified placement constraints. Requests for
HICANNs or NeuronBlocks were already checked. Now requests to place on
logical neurons are checked as well. Therefore the check was moved to a
new place.

Closes bug #3086.

Change-Id: I3e94d65b527d47706f9cd7b9f848500695e4967c