Add note to use of passive repeaters
Passive repeaters will no longer be used in L1-locking tests, cf. c/9503
and c/9605 since the INPUTONLY mode does not exist. If one wants to use
them in future for testing the forwarding mode has to be used.
Furthermore, one has to ensure to not affect other parts of the network
by sending signals in used lines.

Change-Id: Ia30797cf34dd257104b01e58375c657d754a1301