Fix c/10828 (cf. #3538)
In c/10828 blacklisted repeaters were not removed but their attached
buses. This prohibits keeping buses when repeaters are only blacklisted
because of their inherent (mis-)functionality and not because they have an effect
on their attached buses.

With this commit, repeaters and buses need separate blacklisting depending on the reason
of the blacklisting.

Change-Id: I5ad93a16c0248dd43a360b5aed957a18e17f0452